# Refactor: `MovingInterquartileMean`