@Swiftb0y 2024-10-08
So I added yet another optimization (custom `replace_sorted` algorithm instead of `erase`+`insert`) which adds further speedups, especially with realistic IQM sizes.

![MovingIQM Benchmark graph](https://github.com/user-attachments/assets/f704930b-9065-4d55-a8ec-f4a7db4ae559)

Please review.