@Swiftb0y 2024-10-30
I decided to drop `replace_sorted` as it introduced quite a lot of extra code without much benefit. This should now quite drastically simplify the code while still providing a speedup.